Big upgrades towards Linux compatibility - Next step is to move from Objective C to C++ or some other kind of bridge where Swift can reliably

+++ arbitrage-bot/Arbitrage Bot/Arbitrager/include/arbitrager.h
@@ -13,8 +13,6 @@
 #include <stdbool.h>
 #include <stdint.h>
 // Main function to start the server
-
-// Token
 typedef struct {
     const char *name;
     const unsigned char *address; // UInt8
@@ -23,7 +21,7 @@ typedef struct {
 // Server system
 typedef struct {
     void *wrapper;
-    void (*on_tick)(const double* rates, const CToken* tokens, int size);
+    void (*on_tick)(const double* rates, const CToken* tokens, size_t size);
 } PriceDataStore;
 
 typedef struct {
